Support future versions of XJC on a best-effort basis
This removes the hardwired block that prevents the plugin from being used
against any future version of XJC where the MANIFEST.MF Specification-Version
does not match a known version.

This will warn in this situation at this time.

This commit is higher up the patch-set so it can be better tested before
support for 4.0 is added.

I would be my intention to progress this commit into requiring at least a
configuration boolean to be set to authorise this best-effort so that consumers
automated build system start to fail and an administrative decision is made to
proceed.
The decision would be based on an informed error message about what to do and
where to look in documentation for support.