XjcToolFutureIntegrationTest disable IT
Due to 4.0 support being added there is no easily available version
of the XJC tool we can switch to that the plugin does not support.

This is good news for the plugin project, but not this test.

So it has been disabled.

It should be possiblle to validate to test works if you wind back SCM
history to somewhere between the time the IT was added and a commit or
2 before it was disabled.