Issue#38 docLocale switch to using String type
It is not entirely clear to me why using Locale type directly is better
than just keeping the setting as a string and converting it just before
use.

The Gradle documentation does indicate that anything Serializable can be
used as a Property, Locale is Serialiable so I'm not sure why it doesn't
work as-is.

It would be ideal if it was to be kept as Locale type and if both a Locale
or a String can be assigned to the property and it just works.  But I
could not find out how to provide an accepted type convertor that works
in both Groovy and Kotlin DSLs.

The other alternative is to proivide a function like:
  docLocale = locale('it')
  docLocale = locale(Locale.ITALIAN)

It it not clear if the feature previously worked when set inside the DSL,
I was unable to create a TestCase that demonstrated it did.  It only seemed
to work as a gradle.properties setting.

This commit does not change the previous behaviour with regards to
gradle.properties use, but it does allow it to be set from inside the DSL
now in a consistent way to the gradle.properties string property.