asciidoctor bump 2.4.0 to 3.3.2, 2.4.1 to 2.5.13
asciidoctorj-groovy-dsl-1.6.0 does not appear to be on central
maybe it never was, it seems to have disappeared from being available
maybe it was only on jcenter() but this project is not configured
to use jcenter.  Maybe related to CVE-2020-17521 ?

So this bump is the maximum version that still allows JRE8 to be used,
still trying to allow JDK8 to build the project without using a later
JDK with class file compatibility.

+    // going to 4.0.0+ forces JDK11+ for doc production
+    id("org.asciidoctor.jvm.convert") version "3.3.2"

+        // going to 3.0.0+ forces JDK11+ for doc production
+        setVersion("2.5.13")